requirements.txt
hyperliquid-python-sdk>=0.8.0
eth-account>=0.11.0
pandas>=2.0.0
numpy>=1.26.0
schedule>=1.2.0
fastapi>=0.110.0
uvicorn>=0.27.0
